backend/requirements.txt
fastapi
uvicorn
pydantic-settings
python-dotenv
sqlalchemy[asyncio]
asyncpg
langchain
langchain-openai
pgvector
langgraph
langfuse
